+++ src/types.js
@@ -1,5 +1,27 @@
 /* @flow */
 
-export type ParameterType = 'bool' | 'optional-string' | 'required-string' | 'optional-string-variadic' | 'required-string-variadic'
-export type Option = { aliases: Array<string>, parameters: Array<ParameterType>, parameterNames: Array<string>, description: string, defaultValues: Array<any> }
-export type Command = { command: Array<string>, parameters: Array<ParameterType>, parameterNames: Array<string>, description: string, callback: ?((command: string) => void) }
+export type Parameter = {
+  type: 'optional' | 'required' | 'optional-variadic' | 'required-variadic',
+  name: string,
+}
+
+export type Option = {
+  command: ?string,
+  aliases: Array<string>,
+  parameter: ?Parameter,
+  description: string,
+  defaultValue: any,
+}
+
+export type OptionEntry = {
+  name: string,
+  value: any,
+  option: Option,
+}
+
+export type Command = {
+  name: string,
+  parameters: Array<Parameter>,
+  description: string,
+  callback: ?((command: string) => void)
+}